pyproject.toml
[project]
name = "dietagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sse-starlette>=1.6.5",
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "yfinance>=0.2.54",
    "litellm>=1.63.11",
    "json-repair>=0.7.0",
    "jinja2>=3.1.3",
    "duckduckgo-search>=8.0.0",
    "inquirerpy>=0.3.4",
    "arxiv>=2.2.0",
    "mcp>=1.6.0",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph-cli[inmem]>=0.2.10",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "dotenv>=0.9.9",
    "langchain-anthropic>=0.3.16",
    "langchain-qwq>=0.2.0",
    "dashscope>=1.23.6",
    "jupyter>=1.1.1",
]
